+++ packages/medusa/src/api/routes/admin/collections/list-collections.ts
@@ -102,10 +102,11 @@ import { Type } from "class-transformer"
  *     label: cURL
  *     source: |
  *       curl '{backend_url}/admin/collections' \
- *       -H 'Authorization: Bearer {api_token}'
+ *       -H 'x-medusa-access-token: {api_token}'
  * security:
  *   - api_token: []
  *   - cookie_auth: []
+ *   - jwt_token: []
  * tags:
  *   - Product Collections
  * responses: