+++ packages/medusa/src/api/routes/admin/shipping-options/get-shipping-option.ts
@@ -25,10 +25,11 @@ import { defaultFields, defaultRelations } from "."
  *     label: cURL
  *     source: |
  *       curl '{backend_url}/admin/shipping-options/{id}' \
- *       -H 'Authorization: Bearer {api_token}'
+ *       -H 'x-medusa-access-token: {api_token}'
  * security:
  *   - api_token: []
  *   - cookie_auth: []
+ *   - jwt_token: []
  * tags:
  *   - Shipping Options
  * responses: