+++ www/apps/api-reference/specs/admin/code_samples/Shell/admin_order-edits_{id}_items_{item_id}/post.sh
@@ -1,4 +1,4 @@
 curl -X POST '{backend_url}/admin/order-edits/{id}/items/{item_id}' \
--H 'Authorization: Bearer {api_token}' \
+-H 'x-medusa-access-token: {api_token}' \
 -H 'Content-Type: application/json' \
 --data-raw '{ "quantity": 5 }'