+++ www/apps/api-reference/specs/admin/code_samples/Shell/admin_product-categories/post.sh
@@ -1,5 +1,5 @@
 curl -X POST '{backend_url}/admin/product-categories' \
--H 'Authorization: Bearer {api_token}' \
+-H 'x-medusa-access-token: {api_token}' \
 -H 'Content-Type: application/json' \
 --data-raw '{
     "name": "Skinny Jeans"