+++ www/apps/api-reference/specs/admin/code_samples/Shell/admin_regions_{id}_fulfillment-providers_{provider_id}/delete.sh
@@ -1,2 +1,2 @@
 curl -X DELETE '{backend_url}/admin/regions/{id}/fulfillment-providers/{provider_id}' \
--H 'Authorization: Bearer {api_token}'
+-H 'x-medusa-access-token: {api_token}'